Attempting to fix balloon's abuse of CSS positioning.
The balloon container is not where you think it is.  Its size is
100%, but its position is below the Cesium widget, not on top of
the widget.  If it were on top, it would block the mouse from
manipulating anything but the balloon itself.

This change is an attempt to fix, but has the side-effect of making
the already too-tall balloon appear even taller when it is first
activated.  Parking this on a branch for further inspection.